refactor(db): move seed dataset into lib/db/devtools
The dataset is pure data over lib/db schemas; moving it under lib inverts the dependency so the upcoming on-device reseed can consume it (app -> scripts imports are out of bounds, scripts -> lib is fine).
